back-end/requirements.txt
Flask
Flask-Cors
azure-core
azure-ai-textanalytics
azure-search-documents
openai
python-dotenv
